Refactored game play

+++ producer/src/main/java/io/pivotal/producer/square/SquareAccumulator.java
@@ -6,6 +6,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
 import java.util.function.Consumer;
@@ -21,7 +22,8 @@ public class SquareAccumulator {
 
     public List<Square> acquireSquares(UUID gameId) {
         publishPylMessage.accept(createPylMessage(gameId));
-        return accumulateSquares(gameId);
+        LocalDateTime timeout = LocalDateTime.now().plusSeconds(5);
+        return accumulateSquares(gameId, timeout);
     }
 
     private PylMessage createPylMessage(UUID gameId) {
@@ -29,17 +31,17 @@ public class SquareAccumulator {
     }
 
     @SneakyThrows
-    private List<Square> accumulateSquares(UUID gameId) {
-        LocalDateTime timeout = LocalDateTime.now().plusSeconds(5);
-        while (squareRepository.getGameSquares(gameId).size() < 10 && LocalDateTime.now().isBefore(timeout)) {
-            log.info("Found squares {}", squareRepository.getGameSquares(gameId).size());
+    private List<Square> accumulateSquares(UUID gameId, LocalDateTime timeout) {
+        if (squareRepository.getGameSquares(gameId).size() < 10 && LocalDateTime.now().isBefore(timeout)) {
+            Thread.sleep(50);
+            return accumulateSquares(gameId, timeout);
         }
 
         List<Square> gameSquares = squareRepository.getGameSquares(gameId).stream()
                 .map(s -> Square.builder().value(s.getValue()).build())
                 .collect(Collectors.toList());
         squareRepository.removeGameSquares(gameId);
-        return gameSquares;
+        return gameSquares.size() == 10 ? gameSquares : Collections.emptyList();
     }
 
 }
